K8SHelper moved to helper_test.go
