+++ test/e2e/atlas/helper_test.go
@@ -24,9 +24,12 @@ import (
 	"os"
 	"os/exec"
 	"testing"
+	"time"
 
+	"github.com/mongodb-forks/digest"
 	"github.com/mongodb/mongodb-atlas-cli/test/e2e"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 	"go.mongodb.org/atlas/mongodbatlas"
 )
 
@@ -87,6 +90,140 @@ const (
 	e2eMDBVer          = "4.4"
 	e2eSharedMDBVer    = "5.0"
 )
+
+const (
+	defaultTimeout   = 30 * time.Minute
+	defaultTick      = 1 * time.Minute
+	ResourceNotFound = "RESOURCE_NOT_FOUND"
+)
+
+type K8SHelper struct {
+	atlasClient *mongodbatlas.Client
+	project     *mongodbatlas.Project
+	clusters    map[string]*mongodbatlas.Cluster
+	orgID       string
+	t           *testing.T
+}
+
+func NewK8sHelper(t *testing.T) *K8SHelper {
+	t.Helper()
+
+	// These ENV variables MUST be set before running the test
+	orgID := os.Getenv("MCLI_ORG_ID")
+	privateKey := os.Getenv("MCLI_PRIVATE_API_KEY")
+	publicKey := os.Getenv("MCLI_PUBLIC_API_KEY")
+	opsManagerURL := os.Getenv("MCLI_OPS_MANAGER_URL")
+	transport := digest.NewTransport(publicKey, privateKey)
+	tClient, err := transport.Client()
+	require.NoError(t, err, "can not create transport for atlas client")
+
+	atlasClient := mongodbatlas.NewClient(tClient)
+	atlasClient.BaseURL, err = url.Parse(opsManagerURL)
+	require.NoError(t, err, "can not create atlas client")
+
+	return &K8SHelper{
+		atlasClient: atlasClient,
+		t:           t,
+		orgID:       orgID,
+		clusters:    map[string]*mongodbatlas.Cluster{},
+	}
+}
+
+func (kh *K8SHelper) NewProject(project *mongodbatlas.Project) {
+	kh.t.Helper()
+
+	createdProject, resp, err := kh.atlasClient.Projects.Create(context.Background(), project, &mongodbatlas.CreateProjectOptions{})
+	require.Nil(kh.t, err, "error while creating project", err, resp.Status)
+	kh.project = createdProject
+
+	require.Eventually(kh.t, func() bool {
+		_, _, err := kh.atlasClient.Projects.GetOneProject(context.Background(), kh.project.ID)
+		return assert.NoError(kh.t, err, "waiting for project to be created", kh.project.ID)
+	}, defaultTimeout, defaultTick, "project should've been created")
+
+	kh.t.Cleanup(func() {
+		assert.Eventually(kh.t, func() bool {
+			_, err := kh.atlasClient.Projects.Delete(context.Background(), kh.project.ID)
+			if !assert.NoError(kh.t, err, "can not delete test project", kh.project.ID) {
+				return false
+			}
+			_, _, err = kh.atlasClient.Projects.GetOneProject(context.Background(), kh.project.ID)
+			if err != nil {
+				var apiError *mongodbatlas.ErrorResponse
+				if errors.As(err, &apiError) && (apiError.ErrorCode == "GROUP_NOT_FOUND" || apiError.ErrorCode == ResourceNotFound) {
+					return true
+				}
+				return false
+			}
+			return true
+		}, defaultTimeout, defaultTick, "project should've been deleted", project.ID)
+	})
+}
+
+func (kh *K8SHelper) NewCluster(cluster *mongodbatlas.Cluster) {
+	kh.t.Helper()
+
+	require.True(kh.t, kh.project != nil, "can not create cluster without a project")
+	createdCluster, _, err := kh.atlasClient.Clusters.Create(context.Background(), kh.project.ID, cluster)
+	require.Nil(kh.t, err, "error while creating cluster")
+
+	kh.clusters[cluster.Name] = createdCluster
+
+	require.Eventually(kh.t, func() bool {
+		_, _, err := kh.atlasClient.Clusters.Get(context.Background(), kh.project.ID, cluster.Name)
+		return assert.NoError(kh.t, err, "waiting for cluster to be created", cluster.Name)
+	}, defaultTimeout, defaultTick, "cluster should've been created")
+
+	kh.t.Cleanup(func() {
+		assert.Eventually(kh.t, func() bool {
+			_, err := kh.atlasClient.Clusters.Delete(context.Background(), kh.project.ID, cluster.Name)
+			if !assert.NoError(kh.t, err, "can not delete test cluster", cluster.Name) {
+				return false
+			}
+			_, _, err = kh.atlasClient.Clusters.Get(context.Background(), kh.project.ID, cluster.Name)
+			if err != nil {
+				var apiError *mongodbatlas.ErrorResponse
+				if errors.As(err, &apiError) && (apiError.ErrorCode == "CLUSTER_NOT_FOUND" || apiError.ErrorCode == ResourceNotFound) {
+					return true
+				}
+				return false
+			}
+			return true
+		}, defaultTimeout, defaultTick, "cluster should've been deleted", cluster.Name, cluster.ID)
+	})
+}
+
+func (kh *K8SHelper) NewServerlessInstance(instance *mongodbatlas.ServerlessCreateRequestParams) {
+	kh.t.Helper()
+	require.True(kh.t, kh.project != nil, "can not create serverless instance without a project")
+	createdInstance, _, err := kh.atlasClient.ServerlessInstances.Create(context.Background(), kh.project.ID, instance)
+	require.Nil(kh.t, err, "error while creating serverless instance")
+
+	kh.clusters[createdInstance.Name] = createdInstance
+
+	require.Eventually(kh.t, func() bool {
+		_, _, err := kh.atlasClient.ServerlessInstances.Get(context.Background(), kh.project.ID, createdInstance.Name)
+		return assert.NoError(kh.t, err, "waiting for serverless instance to be created", createdInstance.Name)
+	}, defaultTimeout, defaultTick, "serverless instance should've been created")
+
+	kh.t.Cleanup(func() {
+		assert.Eventually(kh.t, func() bool {
+			_, err := kh.atlasClient.ServerlessInstances.Delete(context.Background(), kh.project.ID, createdInstance.Name)
+			if !assert.NoError(kh.t, err, "can not delete test cluster", createdInstance.Name) {
+				return false
+			}
+			_, _, err = kh.atlasClient.ServerlessInstances.Get(context.Background(), kh.project.ID, createdInstance.Name)
+			if err != nil {
+				var apiError *mongodbatlas.ErrorResponse
+				if errors.As(err, &apiError) && (apiError.ErrorCode == "SERVERLESS_INSTANCE_NOT_FOUND" || apiError.ErrorCode == ResourceNotFound) {
+					return true
+				}
+				return false
+			}
+			return true
+		}, defaultTimeout, defaultTick, "serverless instance should've been deleted", createdInstance.Name, createdInstance.ID)
+	})
+}
 
 func deployClusterForProject(projectID string) (string, error) {
 	cliPath, err := e2e.AtlasCLIBin()
